Move SparseArrays support to an extension
